Add startsWith to remote and remoteCached mappers.
Also reworked how requests to the server work, letting people include
ANY character they want in their key strings.

There was a hidden issue where a slash inside of a key would cause the
server to error out, because it was interpreting that as another segment
of the URL, and not finding anything matching that path. The fix was to
switch to using GET data and message bodies for requests.

In theory, there should maybe be a new test added which confirms that
you can use any and every character as the key in the datastore, but
that seems like a lot of work right now.